feat(inquirer): update codebase from inquirer to @inquirer/prompts

+++ src/cli/prompts.ts
@@ -1,22 +1,18 @@
+import { confirm, input, select } from "@inquirer/prompts";
+import chalk from "chalk";
+
 import logger from "../helpers/logger";
 import validate from "./validation";
-import chalk from "chalk";
-import inquirer from "inquirer";
-import SearchList from "inquirer-search-list";
-
-inquirer.registerPrompt("search-list", SearchList);
 
 const selectStarterOrExample = async (): Promise<string> => {
-  const { starter } = await inquirer.prompt({
-    type: "list",
+  return await select({
     message: `How would you like to start your new project?`,
-    name: "starter",
     choices: [
       {
-        name: `Prisma Starter (Recommended)\n\xa0\xa0- ${chalk.gray(
-          "The Prisma Starter is pre-configured with Prisma ORM, Accelerate, Pulse, and a Prisma Postgres database.",
-        )}`,
+        name: `Prisma Starter (Recommended)`,
         value: "starter",
+        description:
+          "The Prisma Starter is pre-configured with Prisma ORM, Accelerate, Pulse, and a Prisma Postgres database.",
       },
       {
         name: "Explore other examples",
@@ -24,101 +20,74 @@ const selectStarterOrExample = async (): Promise<string> => {
       },
     ],
   });
-  return starter;
 };
 
-const getTemplate = async (projects: string[]): Promise<string> => {
+const getTemplate = async (projects: string[]) => {
   logger.success(
     `\nDon't see what you're looking for? Request a new template here:\n\xa0\xa0➡ ${chalk.underline.gray(
       "https://pris.ly/prisma-examples-suggestion",
     )}\n`,
   );
-  const { template } = await inquirer.prompt({
-    // @ts-expect-error Inquirer doesn't register the type.
-    type: "search-list",
+  const result = await select({
     message: `Which template would you like to use?`,
-    name: "template",
     choices: projects,
-    validate: (answer) => {
-      try {
-        validate.project(projects, answer);
-      } catch (e) {
-        return false;
-      }
-      return true;
-    },
   });
 
-  return template;
+  return result;
 };
 
-const selectORMorPDP = async (): Promise<string> => {
+const selectORMorPDP = async () => {
   logger.success(
     `\nThese options correspond to the root directories in the prisma-examples repository:\n`,
   );
-  const { start } = await inquirer.prompt({
-    type: "list",
+  return await select({
     message: `Which Prisma examples would you like to explore?`,
-    name: "start",
     choices: [
       {
-        name: `Prisma ORM\n\xa0\xa0- ${chalk.gray(
-          "Define Prisma schema and run queries",
-        )}`,
+        name: `Prisma ORM`,
         value: "orm",
+        description: "Define Prisma schema and run queries",
       },
       {
-        name: `Prisma Accelerate\n\xa0\xa0- ${chalk.gray(
-          "Perform caching and connection pooling",
-        )}`,
+        name: `Prisma Accelerate`,
         value: "accelerate",
+        description: "Perform caching and connection pooling",
       },
       {
-        name: `Prisma Pulse\n\xa0\xa0- ${chalk.gray(
-          "Monitor and react to real-time database changes",
-        )}`,
+        name: `Prisma Pulse`,
         value: "pulse",
+        description: "Monitor and react to real-time database changes",
       },
       {
-        name: `Prisma Optimize\n\xa0\xa0- ${chalk.gray(
-          "Analyze and improve query performance",
-        )}`,
+        name: `Prisma Optimize`,
         value: "optimize",
+        description: "Analyze and improve query performance",
       },
     ],
   });
-
-  return start;
 };
 
-const getInstallSelection = async (): Promise<boolean> => {
-  const { packages } = await inquirer.prompt({
-    type: "confirm",
+const getInstallSelection = async () => {
+  return confirm({
     message: `Should we automatically install packages for you?`,
-    name: "packages",
     default: true,
   });
-  return Boolean(packages);
 };
 
 const selectManager = async (): Promise<"npm" | "yarn" | "pnpm"> => {
-  const { manager } = await inquirer.prompt({
-    type: "list",
+  return await select({
     message: "Which package manager do you prefer?",
-    name: "manager",
     default: process.env.npm_config_user_agent,
     choices: ["npm", "yarn", "pnpm"],
   });
-  return manager;
 };
 
 const getProjectName = async (defaultValue = ""): Promise<string> => {
-  const { dirname } = await inquirer.prompt({
-    type: "input",
+  return await input({
     message: "What should the project folder be named?",
-    name: "dirname",
     default: defaultValue,
-    filter: (input) => input.replace("/", "_").trim(),
+    transformer: (input) => input.replace("/", "_").trim(),
+    // filter: (input) => input.replace("/", "_").trim(),
     validate(answer) {
       try {
         validate.directoryName(answer);
@@ -129,18 +98,12 @@ const getProjectName = async (defaultValue = ""): Promise<string> => {
       return true;
     },
   });
-  return dirname;
-};
-
-const getProjectDirectory = async (): Promise<string> => {
-  return ".";
 };
 
 export default {
   selectStarterOrExample,
   selectORMorPDP,
   getInstallSelection,
-  getProjectDirectory,
   getProjectName,
   selectManager,
   getTemplate,
